+++ software/nbia-download-manager/src/gov/nih/nci/nbia/StandaloneDMDispatcher.java
@@ -571,7 +571,7 @@ public class StandaloneDMDispatcher {
 
 			HttpParams httpParams = new BasicHttpParams();
 			HttpConnectionParams.setConnectionTimeout(httpParams, 500000);
-			HttpConnectionParams.setSoTimeout(httpParams, new Integer(120000));
+			HttpConnectionParams.setSoTimeout(httpParams, new Integer(720000));
 			httpClient = new DefaultHttpClient(ccm, httpParams);
 			httpClient.setRoutePlanner(new ProxySelectorRoutePlanner(schemeRegistry, ProxySelector.getDefault()));
 			// // Additions by lrt for tcia -
