+++ software/nbia-download-manager/src/gov/nih/nci/nbia/StandaloneDMV3.java
@@ -372,7 +372,7 @@ public class StandaloneDMV3 extends StandaloneDM {
 			// HttpConnectionParams.setSoTimeout(httpParams, new
 			// Integer(12000));
 			HttpConnectionParams.setConnectionTimeout(httpParams, 500000);
-			HttpConnectionParams.setSoTimeout(httpParams, new Integer(120000));
+			HttpConnectionParams.setSoTimeout(httpParams, new Integer(720000));
 			httpClient = new DefaultHttpClient(ccm, httpParams);
 			httpClient.setRoutePlanner(new ProxySelectorRoutePlanner(schemeRegistry, ProxySelector.getDefault()));
 			// // Additions by lrt for tcia -
